requirements.txt
aniso8601==8.0.0
ansible==5.6.0
ansible-core==2.12.4
attrs==19.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.1
configparser==5.2.0
cryptography==3.4.7
cycler==0.10.0
dataclasses-json==0.5.7
decorator==4.4.2
Deprecated==1.2.13
distlib==0.3.4
distro==1.5.0
dnspython==2.2.1
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
erpnext==14.0.0.dev0
filelock==3.6.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Simple==0.0.3
Flask-Mail==0.9.1
Flask-MySQL==1.5.1
Flask-PyMongo==2.3.0
flask-restplus==0.11.0
frappe-bench==5.8.1
gitdb2==2.0.6
GitPython==2.1.15
gocardless-pro==1.22.0
googlemaps==4.6.0
gunicorn==20.1.0
hiredis==2.0.0
honcho==1.1.0
idna==2.10
importlib-metadata==1.5.0
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonobject==2.0.0
jsonschema==3.2.0
kiwisolver==1.3.1
linecache2==1.0.0
MarkupSafe==1.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.22
networkx==2.5.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
paramiko==2.7.2
passwordgenerator==1.4
Pillow==8.2.0
plaid-python==7.2.1
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.14.0
psutil==5.8.0
py==1.11.0
pycountry==20.7.3
pycparser==2.21
PyGithub==1.55
PyJWT==1.6.4
pymongo==4.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.15.7
PySocks==1.7.1
pytesseract==0.3.9
pytest==7.1.1
pytest-ansible==2.2.4
python-crontab==2.4.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-stdnum==1.17
python-youtube==0.8.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.4.1
redis==3.5.3
redisearch==2.1.1
rejson==0.5.6
requests==2.25.1
requests-oauthlib==1.3.1
resolvelib==0.5.4
semantic-version==2.8.5
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
taxjar==1.9.2
testfm==0.1.4
texttable==1.6.3
tomli==2.0.1
traceback2==1.4.0
tweepy==3.10.0
typing-extensions==4.1.1
typing-inspect==0.7.1
Unidecode==1.2.0
unittest2==1.1.0
urllib3==1.26.9
virtualenv==20.14.0
websocket-client==1.3.2
Werkzeug==0.14.1
wrapt==1.14.0
zipp==3.8.0
